Restructured ECS directory
