+++ src/ECS/EntityManager.hpp
@@ -6,25 +6,7 @@
 #include <bitset>
 #include <cassert>
 
-//this will be my ECS system
-// https://medium.com/@savas/nomad-game-engine-part-2-ecs-9132829188e5 
-//this is great because it has nice pictures 
-//https://austinmorlan.com/posts/entity_component_system/
-//components just hold data. Systems hold logic.
-//std::bitset
-//This is cool ^, real cool 
-
-//im going to replace the current system with one closely based off of austins exaple. Because its nice and I can understand it :)
-//austin suggests using 2 maps, one entity/key pair, and a key/entity pair. why? no clue
-
-//austin uses a type template to register types as components
-
-//using a std::set to store keys for each system to iterate through
-
-//my asset handler class will be my version of the coordinator
-//but will include loading and saving things to disk (Ex. gltf stuff, editor things)
-
-namespace engine {
+namespace ECS {
 
 using Entity = std::uint32_t;
 static const Entity MAX_ENTITIES = 5000; //#def const? what diff? If I use a vector I dont need this constant
@@ -33,7 +15,8 @@ using ComponentType = std::uint8_t;
 const ComponentType MAX_COMPONENTS = 32;
 
 using Signature = std::bitset<MAX_COMPONENTS>; // used to determine which components an entity has
-
+    //Entities are the ID's that accociate components together
+    
     class EntityManager {
     public:
         EntityManager() {
@@ -53,6 +36,7 @@ using Signature = std::bitset<MAX_COMPONENTS>; // used to determine which compon
             return id;
         };
 
+        //cleaning up stale component data of a deleted entity is handled elsewhere
         void destroyEntity(Entity entity) {
             assert(entity < MAX_ENTITIES && "Entity can never have existed");
 
@@ -61,6 +45,17 @@ using Signature = std::bitset<MAX_COMPONENTS>; // used to determine which compon
             //keeps data density by pushing!!!
             availableEntities.push(entity);
             --livingEntities;
+        }
+
+        void SetSignature(Entity entity, Signature signature) {
+            assert(entity < MAX_ENTITIES && "Entity can never have existed");
+
+            signatures[entity] = signature;
+        }
+        Signature GetSignature(Entity entity) {
+            assert(entity < MAX_ENTITIES && "Entity can never have existed");
+
+            return signatures[entity];
         }
 
     private:
